fix: Disabling mocks during testing
